implemeting applyTranslations()
not yet working perfectly
